+++ lib/PayPal/Rest/OauthHandler.php
@@ -1,14 +1,12 @@
 <?php
 /**
- * API handler for all REST API calls
+ * API handler for OAuth Token Request REST API calls
  */
 
 namespace PayPal\Rest;
 
-use PayPal\Auth\OAuthTokenCredential;
 use PayPal\Common\PPUserAgent;
 use PayPal\Core\PPConstants;
-use PayPal\Core\PPCredentialManager;
 use PayPal\Core\PPHttpConfig;
 use PayPal\Exception\PPConfigurationException;
 use PayPal\Exception\PPInvalidCredentialException;
@@ -16,9 +14,9 @@ use PayPal\Exception\PPMissingCredentialException;
 use PayPal\Handler\IPPHandler;
 
 /**
- * Class RestHandler
+ * Class OauthHandler
  */
-class RestHandler implements IPPHandler
+class OauthHandler implements IPPHandler
 {
     /**
      * Private Variable
@@ -48,42 +46,20 @@ class RestHandler implements IPPHandler
      */
     public function handle($httpConfig, $request, $options)
     {
-
-        $credential = $this->apiContext->getCredential();
         $config = $this->apiContext->getConfig();
-
-        if ($credential == null) {
-            // Try picking credentials from the config file
-            $credMgr = PPCredentialManager::getInstance($config);
-            $credValues = $credMgr->getCredentialObject();
-
-            if (!is_array($credValues)) {
-                throw new PPMissingCredentialException("Empty or invalid credentials passed");
-            }
-
-            $credential = new OAuthTokenCredential($credValues['clientId'], $credValues['clientSecret']);
-        }
-
-        if ($credential == null || !($credential instanceof OAuthTokenCredential)) {
-            throw new PPInvalidCredentialException("Invalid credentials passed");
-        }
 
         $httpConfig->setUrl(
             rtrim(trim($this->_getEndpoint($config)), '/') .
             (isset($options['path']) ? $options['path'] : '')
         );
 
-        if (!array_key_exists("User-Agent", $httpConfig->getHeaders())) {
-            $httpConfig->addHeader("User-Agent", PPUserAgent::getValue(PPConstants::SDK_NAME, PPConstants::SDK_VERSION));
-        }
+        $headers = array(
+            "User-Agent"    => PPUserAgent::getValue(PPConstants::SDK_NAME, PPConstants::SDK_VERSION),
+            "Authorization" => "Basic " . base64_encode($options['clientId'] . ":" . $options['clientSecret']),
+            "Accept"        => "*/*"
+        );
+        $httpConfig->setHeaders($headers);
 
-        if (!is_null($credential) && $credential instanceof OAuthTokenCredential && is_null($httpConfig->getHeader('Authorization'))) {
-            $httpConfig->addHeader('Authorization', "Bearer " . $credential->getAccessToken($config), false);
-        }
-
-        if ($httpConfig->getMethod() == 'POST' || $httpConfig->getMethod() == 'PUT') {
-            $httpConfig->addHeader('PayPal-Request-Id', $this->apiContext->getRequestId());
-        }
         // Add any additional Headers that they may have provided
         $headers = $this->apiContext->getRequestHeaders();
         foreach ($headers as $key => $value) {
@@ -92,33 +68,38 @@ class RestHandler implements IPPHandler
     }
 
     /**
-     * End Point
+     * Get HttpConfiguration object for OAuth API
      *
      * @param array $config
      *
-     * @return string
+     * @return PPHttpConfig
      * @throws \PayPal\Exception\PPConfigurationException
      */
-    private function _getEndpoint($config)
+    private static function _getEndpoint($config)
     {
-        if (isset($config['service.EndPoint'])) {
-            return $config['service.EndPoint'];
+        if (isset($config['oauth.EndPoint'])) {
+            $baseEndpoint = $config['oauth.EndPoint'];
+        } else if (isset($config['service.EndPoint'])) {
+            $baseEndpoint = $config['service.EndPoint'];
         } else if (isset($config['mode'])) {
             switch (strtoupper($config['mode'])) {
                 case 'SANDBOX':
-                    return PPConstants::REST_SANDBOX_ENDPOINT;
+                    $baseEndpoint = PPConstants::REST_SANDBOX_ENDPOINT;
                     break;
                 case 'LIVE':
-                    return PPConstants::REST_LIVE_ENDPOINT;
+                    $baseEndpoint = PPConstants::REST_LIVE_ENDPOINT;
                     break;
                 default:
                     throw new PPConfigurationException('The mode config parameter must be set to either sandbox/live');
-                    break;
             }
         } else {
             throw new PPConfigurationException(
                 'You must set one of service.endpoint or mode parameters in your configuration'
             );
         }
+
+        $baseEndpoint = rtrim(trim($baseEndpoint), '/') . "/v1/oauth2/token";
+
+        return $baseEndpoint;
     }
 }